Rename: change file names and move to src folder
