+++ main.js
@@ -4,14 +4,14 @@ import path from "path";
 import chalk from "chalk";
 import boxen from "boxen";
 import figlet from "figlet";
-import { loadConfig } from "./configLoader.js";
-import { createConfig } from "./create-config.js";
-import { renderResult } from "./result.js";
-import { getTailwindCss } from "./tailwindCss.js";
-import { changeToPolyfill } from "./polyfill.js";
-import { checkCssCompatibility } from "./cssCompatibility.js";
-import { getStyledComponentsCss } from "./styledComponents.js";
-import { selectBrowsersAndVersions } from "./userSelection.js";
+import { loadConfig } from "./src/configLoader.js";
+import { createConfig } from "./src/create-config.js";
+import { renderResult } from "./src/result.js";
+import { getTailwindCss } from "./src/tailwindCss.js";
+import { changeToPolyfill } from "./src/polyfill.js";
+import { checkCssCompatibility } from "./src/cssCompatibility.js";
+import { getStyledComponentsCss } from "./src/styledComponents.js";
+import { selectBrowsersAndVersions } from "./src/userSelection.js";
 
 const currentPath = process.cwd();
 const args = process.argv.slice(2);
