+++ src/resultGenerator.js
